# post_service/storage/repo/models.go
package repo

type Post struct {
	Id          string
	UserId      string
	Title       string
	Description string
	Likes       int64
	UserName    string
	Comments    int64
	CreatedAt   string
	UpdatedAt   string
}